backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.16.0
python-dotenv==1.1.1
pydantic[email]==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 